Use cached policies container for speed tagging
This behaves slightly differently now: prior to this change, any
policies already listed as check boxes were excluded from the
supplimentary select field. This is not entirely necessary and also
makes it harder to use the cached data. Besides which, this field is
rarely used during the speed tagging process. So instead we list all
policies and make sure any that are already listed as checkboxes are not
selected to avoid them appearing twice when checked.