Add row wrapping div to fix IE7 float bug
IE7 doesn't quite understand floating and clearing elements so doesn't
position the headers next to the list of links they are associated to.
Instead the headers float up to the list above the one it is intended
for.

This adds a wrapping div to each list of links which forces the heading
to clear the previous list and sit next to the list it is intended for.