Gracefully handle re-editioning during interim
This is some more temporary code to handle the state between the code being
deployed and the data migration to move the body to the new model has completed.
If a user should re-edition a document during this period where a HtmlAttachment
doesn't have an associated GovspeakContent instance (i.e. before the data
migration has completed), then we want to clone the attachment and manually
create the new assocation. This will be removed in the cleanup pull request to
come later.