Fix add_content_id_to_orgs migration
With the migration added in ba5a421 there was both a `null:false` and
index on the new `content_id` column. This prevented the migration from
running as when it came to add the index all the rows in the column were
blank strings.

This commit changes the existing migration to add the `content_id`
column without the null constraint and updates the data migration to
look for Orgs with nil `content_id` fields instead of empty strings.