Ensure namespaced class is findable
We get occasional uninitialized constant exceptions because the
Whitehall::ExtraQuoteRemover class cannot be found. This adds the class to its
parent module such that it can be found regardless of the execution path
of the code. See http://urbanautomaton.com/blog/2013/08/27/rails-autoloading-hell/
for more details about constant lookup and Rails autoload.