Rename publishing API workers
Neither "register" nor "artefact" are terms that are appropriate for the new
publishing API. This renames the workers to better reflect what they are for.

Note that the old files are left in place with the constants set as temporary
aliases of the new classes. This is so that any jobs already queued against the
old worker classes can still be processed. This will need to be cleaned up in a
future commit.