Add support to filter announcements by date
Can ask for past or future announcements only. When asking for future statistics
announcements, they are returned in date order, i.e. the closest one first. This
makes it more likely that users will see the announcements that are most
pertinent, i.e. those that are due soon but do not yet have a linked
publication.