Ensure announcements scope loads the right things
So this is quite complicated... It's an example of a "greatest n by group"
query, but on a join model that we are eager loading. We want to get all the
statistics announcements for the given criteria, ordering them by the release
date on the most recent joined statistics announcement date.