Remove unnecessary indirection
The use of a separate presenter here on the statistics_announcements#show action
is unnecessary indirection as pretty much all the methods required in the view
are already present on the model. It also makes it more likely that bugs like
the one fixed by the previous commit will occur.