Prevent test from failing due to timing issue
The test here is not concerned with the order these editions are returned in,
just the fact that the correct ones are returned. By using assert_same_elements,
we don't have to worry about the nondeterministic order they may be returned in
because they are created at the same time (time is frozen in the test suite).