Change Ofqual slug to match search term
Ofqual was initially created as the unabbreviated version but the title
is Ofqual, and that's what people think of it as and search for.

This commit adds a data migration which updates the slug, along with all
of Ofquals Users and Editions which they've published.

This will also require a bunch of changes in Production environments
which I don't have access to.

Ticket: https://www.pivotaltracker.com/story/show/80284724