Collection cache max age should be 5m
www.agileplannerapp.com/boards/173808/cards/8023

HMRC need to publish a collection page which needs
to reflect the latest changes in no longer than 5m.

made a change to the `CacheControlHelper` to allow
overriding the `cache_max_age` from the controller,
so that we can set a max age of 5m for collections.