Publishing API: Include an image for case studies
If the case study or it's org has an image available, we include it in the JSON
sent to the Publishing API.

Note that this code will not fall back to the "placeholder" image if neither the
case study or the organisation doesn't have an image. This is intentional as the
responsibility for deciding what to display if no image is available should lie
with the public-facing application that renders the content.