Document filter message for no results
www.agileplannerapp.com/boards/173808/cards/8672
addresses: https://govuk.zendesk.com/tickets/869065

In its current version the document filter shows
a loader image when the search is being performed
and doesn't provide any feedback if results are
not found.

Editors can get confused if they're searching for
a publication to link to a Statistics Announcement
and no results are returned. It may seem as if the
filter is not functioning.

I've added a suitable message which reminds editors
that the lookup is scoped by publication type which
is same as the type of the statistics announcement,
and, publication state being not published.