Fix expiry for collections with nothing scheduled
DocumentCollections with none of its items scheduled
for publishing were reporting a max-age of 15.mins,
in spite of [this change](https://github.com/alphagov/whitehall/pull/1837).

That's because expire_on_next_scheduled_publication
doesn't change the expires header if nothing in the
collection is scheduled to publish.

I've overriden the set_cache_control_headers in this
controller to use documents_collection_cache_max_age
which fixes this.