Add change_history to case study API presenter.
This creates some duplication since the top change note within the
change_history hash will be the same as the existing change_note
attribute. However, the latter is already being used by the email alerts
service; it would probably be good to change that app to use
change_history instead.