Fix fatality notice casualty field duplication
www.agileplannerapp.com/boards/173808/cards/8196

this issue seems to have its roots in this refactoring:
https://github.com/alphagov/whitehall/commit/3b8145

the remove button handler script adds an additional
hidden field '_destroy' to the form with value true,
which should be sent to the server, which was not
the case. instead this extra hidden field was being
sent to the server.