Data migration to ensure documents have a content_id
it's better to do this in a data migration so that we are sure that the
application server code has reloaded when this is run.

If this was done in a database migration then there would be a race
condition where:

* all documents had had a content_id set
* application code had not reloaded
* user creates a new draft document which does not get the content_id
  set

We then end up with a few documents without content_ids.