Options passed to the API worker should pass to the presenter
We want to be able to 'republish' documents, so we need to be able to
pass an `update_type` option to the worker.

In an ideal world the test case for this could have been neater. But the
helpers provided by gds-api-adapters do not allow stubbing requests
without doing an exact match on the body json, so we need to
render the entire presenter output.