Fix bug with supporting page url generation
There was an issue with generating URLs for supporting pages that belong
to more than one policy.  The code in the controller[1] specifies the
policy to generate the URL for.  However, this option was being
overridden in the url helper.

This updates the URL helper to only set the policy if one hasn't been
given.

[1] https://github.com/alphagov/whitehall/blob/build-11392/app/controllers/supporting_pages_controller.rb#L10-L11