Avoid serialization issues when invoking PublishingApiWorker
sidekiq serializes all of its arguments and de-serializes them in the
background worker process. This means that symbol-keys of hashes get
converted into strings. This caused a bug because the `update_type`
option was ignored because it got converted from a symbol to a string
key in the options hash.

This commit also corrects the `PublishingApiRepublisherTest` to set the
`update_type` option passed to the presenter when generating the
expected message body for stubs. Previously the tests were passing even
though the code was doing the wrong thing (ignoring the update_type!).