Publish all translations to the Publishing API
Whitehall::PublishingApi.publish and Whitehall::PublishingApi.republish now send
all available translations to the Publishing API.

Note that these methods will now blow up if handed a non-translatable model, but
as we are not yet sending any non-translatable models to the Publishing API,
it's impossible to write a test to support this behaviour. That can be done at
the appropriate time.