Add more links to services and information pages
This commit turns on the link for the HMRC services and information
page: this is a time-critical change which needs to go out on Wednesday
10th December - and HMRC will need to be informed so that they can
remove a duplicate link from their organisation page when it has been
deployed.

It also adds links for several other organisations (DfE, HS2, DVSA);
these ones are not time-critical.