Update policy data migration to use prefix route
Use a prefix redirect for the policy change. This will then match all
the supporting pages and redirect them.

Also remove the 'for' from the search index calls. As far as I can tell
the search index gets set based on the object anyway. It was also
causing the migration to fail for me locally.