Show unpublishing information on drafts
www.agileplannerapp.com/boards/173808/cards/8902

content designers need to know when a draft is a
fresh draft as opposed to a draft that resulted
from an unpublishing.

this is a temporary fix whilst we create a state
called 'unpublished' to represent such editions.

Add reindex edition by slug endpoint

Story: https://www.pivotaltracker.com/story/show/84836766

This adds an endpoint for triggering a reindex of all documents tagged
to a given slug.

Slugs are only unique across document types, so to insure that a
document with the passed in slug gets reindexed, we select across all
types known to the Document model.

Remove unused Document.document_types method