Turn scoped search on for the CMA
Tara's analysis indicates that it's likely to be an improvement to turn
scoped search on for the CMA, though it's not as clear as for other
organisations we've turned this on for.  We'll watch this closely; it
may be that we need to filter to a group of organisations, rather than
just to CMA content.