TL;DR: Don't use refute with Capybara finders
Capybara's finders are written to assume that the thing you are checking for is
present on the page, and so will keep looking for it until it reaches it's
default timeout (which is currently set to 5 seconds). Therefor any use of a
"presence" finder (e.g. has_css?) to prove something is not on the page will add
a hefty delay in the running feature. Instead, the corresponding "absence"
finder (has_no_css?) can be used to avoid the delay.