Update to Ruby 2.1.4
This requires a couple of minor changes:

1. The debugger gem is only compatible with 1.9.X. The byebug gem exists and is
compatible with 2.1.X, but I figured it's better to exclude that until someone
needs it.

2. Setting the default_url_options on the singleton class was causing issues in
the test suite. This is not actually necessary as ActiveSupport class
attributes (which default_url_options is) already ensure that overriding on the
instance doesn't interfere with the parent class' value. Also removes some
unnecessary tests (other tests will cover the inclusion of the url helpers and
also it is very unlikely that they will be removed from the class)