Fix validations on join models
The upgrade to Rails 4 broke the validations on join models when creating new
joins on new records. This is because the join model is now validated before the
records it is joining are validated and saved, which means validating for the
presence of the joining records will fail because they don't yet exist in the
database. We can avoid this issue by updating the relevant associations to
include the inverse association, which means that the in-memory instances get
assigned to the join model and the presence check succeeds without having to try
the database. To get this to work, some of the validations needed adjusting such
that it checks for the presence of the model itself and not the id attribute.