Only show services-info examples from the org
On the services and information pages for an organisation, we were
previously showing the most popular examples for each sub-topic
regardless of whether they were tagged to the organisation.  Instead,
only show examples which are tagged to the organisation.  This helps on
pages like the HMRC services and information page, where some sub-topics
(such as "Maritime safety") were showing no examples which were to do
with HMRC.