No need to explicitly save CIP titles as blank
The with_translations scope is primarily meant for loading translations of a
particular locale, but it is often used in whitehall as a way of pre-fetching
the translations to avoid n+1 queries. The version of the with_translations
method in pre-Rails 4 globalize used to invoke with_required_attributes as well,
which had the effect of excluding any instances that didn't have all required
attributes set in the database. The effect of this was that calling
with_translations on CorporateInformationPages would return an empty set,
because they do not have an explicit title saved in the database as their titles
are derived from their CorporateInformationPageType. We got around this by
setting their titles to a blank string in the database (rather than leaving
them to be NULL). Now, in Rails 4-globalize, the with_translations scope no
longer calls `with_required_attributes` (the method itself is also deprecated),
making the little dance redundant.

I've kept the test in here as a regression, although rewritten it slightly to
focus on the important thing - that the scope still loads the instances.

 # Please enter the commit message for your changes. Lines starting