Different email signup for MHRA
MHRA is a special case in that Users have a strong connection with the
brand so it needs a different URL for the signup page so they can also
sign up for drug safety updates or drug and device alerts.

I've done this by calling a helper from the feed and passing in the
atom_url which is in turn passed to EmailSignupPagesFinder which does a
lookup in an array of OpenStructs to find a matching regex. This also
includes the Rails Route helpers in order to build the path in a sane
way.

Ticket: https://trello.com/c/nmljEkL1/508-mhra-email-sign-up-routes