Don't inherit Decorator from Struct
This was causing Rail's content_tag_for helper to blow up because (as of Rails
4), that method wraps the passed in record in an array and iterates over it (so
that multiple records can be passed in and handled). The problem is that a
Struct is only meant to be a simple data structure, and as such, is enumerable
(it enumerates over its attributes).

This worked with Rails 3 because that version of the helper called `to_ary` on
the passed in record(s) only if it responded to it instead of wrapping in an
array and iterating.

Besides all this, inheriting from Struct is a bit of an anti-pattern[1],
especially for anything more complex than a simple data container.

[1] http://blog.steveklabnik.com/posts/2012-09-01-random-ruby-tricks--struct-new