Refactor custom email signup page linking
This reverts EmailSignupPagesFinder back to its previous interface, adding one
additional method to identify if an email signup page exists for a given atom
feed URL. It moves the path generation to the helper, removing the need to mix
in the routing helpers to the EmailSigupPagesFinder, and also makes that changes
to EmailSignupInformationController unnecessary.

At present EmailSignupPagesFinder only matches the MHRA organisation.