Use patched globalize gem
A change made to the globalize gem that made the belongs_to association
on the Translation model "touch" the owning translated model[1]. This causes
issues when trying to destroy translated instances that are using Rails'
optimistic locking: When destroying a translated instance, the associated
translations are destroyed first, which "touch"es the owning instance
(incrementing the lock_version), making the in-memory lock_version out-of-date,
resulting in the destroy failing with a ActiveRecord::StaleObject error.

This is ultimately a bug in ActiveRecord (it should be able to cope with such a
configuration), but fixing it is non-trivial. For now use a forked version of
globalize that does not touch the association, as this isn't needed
in whitehall (it was introduced to invalidate timestamp-based cache keys when a
translation is destroyed on a model).

[1]: https://github.com/globalize/globalize/pull/330