Fix attachment ordering on new records
The code that set the ordering on new attachments fails when creating a brand
new attachable thing with more than one attachment. This appears to be because
the query generated to calculate the maximum available ordering is not correctly
setting the newly generated ID of the attachable thing, and thus no matching
attachments are being found. The fix is to explicitly generate the query to get
the maximum ordering, making sure the ID of the attachable thing is used.