Reduce duplication in SCSS typography mixins
These share a lot of logic/styling with the typography mixins
in the govuk_frontend_toolkit. FET mixins are the authorative
definition of type styling, and also provide additionalbehaviour
like:
- Switching to a non-webfont for printing, to avoid webfont
  printing issues
- Any future improvements that may be added

This brings WH styling slightly more in line with other
frontends.

This does change some styling though, a couple of the
640px breakpoint sizes are different in WH to FET:

- core-48 @ 640: 25px/0.8 => 53px/1
- core-27 @ 640: 19px/1.3 => 20/1.1
- core-24 @ 640: 17px/1.5 => 18px/1.2
- core-19 @ 640: 13px/1.5 => 16px/1.25
- core-16 @ 640: 13px/1.5 => 14px/1.1

After discussion with @edds we think a ship-it and see
approach is probably safe, as this will only affect
some minor cases