Propogate updates to unpublishing to publishingAPI
https://trello.com/c/H7tPyTil

Users are able to edit the "explanation" on an unpublishing
outside of normal document workflow. At present, these edits
do not get reflected in the "unpublishing" in content store.

Whitehall needs to be updated such that when the explanation
on an unpublishing is made, the "unpublishing" in content
store is also updated.