Replace "locale" column with "primary_locale"
The fact that we are using a column called "locale" on the editions table to
signify the locale for non-english editions causes an issue with later
versions of the globalize gem.

This is the second phase of migrating this column to a new name of
"primary_locale". We pre-emptively added the column to the table in
3f2a376ea75380098fb629e9c03f38359d4a3f4e. This commit includes a
migration to sync the `primary_locale` column with `locale` on
deployment.

At this stage we do not yet remove the old locale column, to avoid
breaking existing running processes which expect the `locale` field to
be present still. We deprecate the `locale` column in the `Edition`
model however.

Once this code has been deployed, the pending migration can be run to
safely drop the old column and the `deprecated_columns` directive can be
removed.