Add political flag to organisations
So that we can correctly flag content which may contain content which is
not the opinion of the current government, we need to be able to tell if
an organisation publishes content associated to the current government.
This adds a flag to organisations which can manually be set in the admin
by GDS editors.

Added a basic test to assert that the checkbox doesn't get displayed for
other users.