Avoid 'Cannot redirect to a parameter hash' error
Rails 4 introduced a [protection against passing parameters hash into
`redirect_to`](https://github.com/rails/rails/pull/16170) as a security
measure.

In some circumstances the `session[:document_filters]` could contain a
paramters hash (I suspect this may only be the case for a session which
was serialized into someones session cookie in the pre-rails-4 whitehall
app and deserialized in this release).

Adding this `.to_h` forces the hash to always be converted to a ruby
hash.

It's safe to pass these parameters to redirect_to because they are all
features which are intended to be controlled by user input. The redirect
is just a feature to ensure consistency of urls.