Preemptively add primary_locale to editions
this is to avoid a race condition when deploying the rails 4 upgrade.

I noticed errors when deploying to preview which I think occured because the
whitehall-frontend apps restarted before the schema change had propogated to
the slave. So the frontend apps saw the old schema and thought that Editions
did not have a `primary_locale` attribute.

We can pre-emptively add and populate this field.

A migration included in the rails 4 branch will update the field to ensure that
it contains the latest data at the time of switchover. Adding it pre-emptively
to the database schema avoids the race condition.