Allow governments to be displayed without end date
An end date isn't needed for a government to be valid. Calling
`.to_s(:govuk_date)` on a `nil` object throws an error. This adds a
check to make sure end date exists before trying to format it as a
`govuk_date`.

Also updated the cucumbers to first create a government with no end
date.