Remove metadata link tracking
As part of moving analytics from classic to universal, we’ve audited
the custom variables being set and whether they are still being used.

Ash identified this variable as being unused and marked it for
deletion. It will not be switched over to Universal.

Undoes 7583621517ed47157b3f29c8a42c835443561bb6