Publish Ministerial Roles to PublishingAPI
This commit adds the ability for Ministerial Roles to be sent to the
PublishingAPI as placeholders. This work is required for polices to
exist in the Content Store. Ticket:
https://trello.com/c/IZDLvzcc/33-add-missing-whitehall-non-documents-to-the-content-store

It also changes the reference in Organisation in
`PublishingApiPresenters::PlaceholderTest#present` as this was
presumably missed when the code was refactored.