Ensure deprecated columns are inherited
It's important that deprecated columns are inherited because if a column
is deprecated in the base class then it would also need to be deprecated
in any derived class (if using Rails STI).

We have a specific issue because `Policy` deprecates
`published_related_publication_count` and Edition deprecates `locale`.
There were errors being raised inside ActiveRecord in some edge cases
because the `locale` column was not deprecated for `Policy`.

https://trello.com/c/ZUlRft6V/53-rails-4-upgrade-deprecated-columns-helper-error