Bugfix: cannot redirect to params hash on export
When exporting a CSV list of documents there would be an error 'cannot
redirect to params hash'.

The values returned from `params_filters` have a limited set of keys which
are safe to use in redirects. So always convert this to a plain hash.

I have also explicitly adjusted code so that the `session[:document_filter]`
uses string keys exclusively, but passes an options hash with symbol keys to
the edition filter. There was some confusion in both the controller and
functional test about this.

I also removed the redundant stub from the test case of the invalid scenario.
This stub was doing nothing. The invalidity of the filter is triggered by
setting a non-existent author.