Update sprockets and sprockets-rails
Whitehall is experiencing intermittent test failures on parallel builds due to
an issue in the sprockets gem[1]. This is fixed in master, but not in a final
release yet. This updates sprockets to the latest 3.0 beta release to get
around the issue.

Importing a folder of SCSS partials using "/*" isn't working with these versions
of sprockets and sprockets-rails. Being explicit with the order that the CSS is
loaded is better as it leaves no ambiguity with the load order/cascade.

[1] https://github.com/sstephenson/sprockets/pull/545