Update Government on Document during publish
The associated Government should be based on the `first_public_at` date,
which can be edited as part of edition metadata, so the associated
Government needs to be updated when the edition is.

This is done after the Edition itself has been saved, in case the
edition update fails.

This also removes the old implementation/tests, which set the
Government once (and only once) during the first publish