Remove extra test which had intermittent failures
The test was ensuring that world locations grouped by type and those
groups were ordered correctly. The test was sometimes failing due to the
locations within the types being ordered by name. As we weren't defining
names in the creation they would both get the standard factory name.
This caused the locations would come back in an indeterminate order.

I have removed this test and updated the test below which was already
testing that locations were ordered by names within the groups. That
test was already testing all the functionality that the removed test was
testing so there isn't any loss in coverage with this change.