Add a history-mode column to Document export CSVs
This will be used by Departmental Editors to see what content will be
affected by history-mode, eg, content that is 'political' and that
belongs to a previous government will be marked as historical.

This allows Editors to see which of their documents will be marked
as historic when the government changes.