Don't reference versions in admin editions filter
This was added to get around an issue during the Rails 4.1 upgrade that meant
ActiveRecord wasn't able to generate a valid query when trying to eager load the
last_authors on a collection of editions. Unfortunately, this resulted in a
highly inefficient queries that required a lot of writing to disk.

To get around the issue whilst we investigate further, the eager-loading of
last_author is being removed. This makes the page n+1 (where n is 50) as it now
loads each `last_author` with an individual query, but the additional impact to
users is fairly minimal on this already-slow page.

The root problem is that ActiveRecord doesn't seem to be able to eager-load the
`last_author` association correctly. This association is a has_one that is
through another has_one association (`most_recent_version`) that also has an
order clause to determine the precise one.