Fix translated child organisations appearing twice
If a parent organisation (as defined by its relationship to
child_organisations) had a child that was translated, it would appear
on the Organisations#index page once per translation.

The real life result of this bug is that HMCTS, which also has a Welsh
version, appears twice under MoJ.

The applied fix is basically the same as in
ec3979c1308a3f576cfe531d775283bf863d724d.