Delay destroying records of dependent editions
https://trello.com/c/oYy2udbc

They were being destroyed when the depended-upon edition was published
so that we clean-up as soon as they're no longer required. Now, we continue
to hold onto these associations till the time the depended-upon edition
is superseded. There's no side-effect because we replublish dependants
only when the depended-upon edition is published / force-published.

Doing this allows us to cover the case of the depended-upon edition being
unpublished, and then published again, in which case we need to republish
dependent editions. This is better explained by the new test added in:

test/unit/services/listeners/edition_dependencies_test.rb
"unpublishing a depended-upon edition and publishing it again
  should cause dependent editions to be republished"

NOTE: this doesn't cover the case where a subsequent edition of the
depended-upon edition changes the title/slug, leaving an outdated
slug in the dependent edition.