Adjust political identification rules
 Both statistical publications and fatality notices should never be identified
as being political. Once these two are excluded, the only remaining edition
formats are those that cannot be associated with ministers, which means that
they also will never be identified as being political (CaseStudy,
CorporateInformationPage, DocumentCollection, Policy, SupportingPage and
WorldwidePriority. This makes the test for non-political formats always being
political if associated with a minister redundant and thus the format can guard
against the simple check of whether there are any ministers or political orgs.