Add script to fix UUIDs in MySQL replication
The bare `uuid` function call in MySQL isn't replication safe when
performing STATEMENT-level binary replication:

http://dev.mysql.com/doc/refman/5.1/en/replication-rbr-safe-unsafe.html

In a previous migration, we batch-updated several models using this
function to insert `content_id` values. When these migrations replicated
to our backup and read-only slaves, the UUIDs became wrong because of
the replication safety issue.

This means that currently our backups aren't current with production as
the `content_id` values differ.

This commit fixes that by adding a script to be run on the master
machine which:

1. Grabs the `content_id` in Ruby-land, avoiding replication issues
2. Forcing an `UPDATE` statement to run using `update_columns` call
3. Once this has run, the `UPDATE` statements will replicate to the
   slaves containing the correct strings, which will synchronise the
   slave values with the master.

The migrations themselves were future-proofed in
4864e94e58f65900fc2b089f966aad75f74ba6c3.

The command should be run with `rails r` - we chose this path over a
data migration because we don't want this script to run inside a
transaction (it takes over 90 seconds to run).