Fix nondeterministic test
This test was failing intermittently. I think the assertion was relying on the
implementation pulling records from the database in the order in which they were
inserted, but the implementation didn't specify an ordering. Whilst we could
write an assertion that the lines for each TagChange appeared in the right order
relative to each other, but in any order within the overall set of lines, we
feel that the assertion can be dropped as it wasn't adding much value anyway.

Since we no longer cared about the logging output in the test, I removed the
stubbing. This meant that it mingled with the test output. Therefore I added
the logger to mute the output by default, but send it to STDOUT when used in
the Rake task.