Add history mode fields to Edition search index
These fields will be used to mark some search results as being under
history mode, where political content from previous governments will
displayed in the UI as associtated to the government that published
it.

Companion Rummager change, adding new fields:
 - https://github.com/alphagov/rummager/pull/378

**New fields:**
- `is_political`: _boolean_
  - If the content is considered political in nature, relfecting
  views of the government it was published under
- `is_historic`: _boolean_
  - If the content is political and published by a previous
  government, it is considered historic and not relfective of the
  current government and can be marked as such.
- `government_name`: _unsearchable_text_
  - The name of the Government that first published
  this document, eg, "1970 to 1974 Conservative government"

** Notes **

`government_name` is stored as text, rather than an identifier, and
not normalised, but storing the identifier would require either
Rummager or the frontend to perform expansion from identifier to
name and governments are not repsented in content store, yet.

If a government name is edited then content associated to that
government will need to be reindexed for the updated name to be
reflected in search results. This isn't something we expect to happen
very often, and are happy running a manual reindex if required.

When a government is ended all political documents will need to
be re-index, to set `is_historic` in search. This is a once every ~5
year change, and we're happy running a reindex task in this instance.

It's possible the updates for a government endeding could be done as
an async task that happens as a result of the government being ended
in the publishing tool. We'll look at that as follow up work, for now
this gets the right fields into search.

As discussed with @edds and @rboulton this morning.