Simplify Government permissions enforcement
There is only really one level of permissions for governments - either a user
can manage governments or they can't - therefor it keeps things simpler to only
have the one action in the authority rules. If it becomes the case that we
require separate rules for creating and editing at some point in the future,
then the rules can be updated accordingly.