Extract model for representing new-world policies
A simple model to provide a cleaner abstraction for dealing with policies from
the new Publishing 2.0 world. At the minute, this simply wraps the data returned
by content register with the added enhancement of exposing the slug. In future,
this can be further enhanced to do the work around figuring out the related
policies (i.e. exposing the parent policy area when linked to a policy
programme)