Improve history notice visibility on document page
Reduce the length of the copy and make the text larger.

Most users of these pages are specialists, rather than mainstream
users, and need less sign-positing that the content is from a
previous government. We think they'll be more interested in
the particular government.

This allows us to reduce the copy significantly, giving us space
to increase the font-size to match the page title. This should
be more impactful and more visible.

It also breaks the sentance on the government name, meaning the
second line will always lead with the XXXX to YYYY date part of
the government, which we think will be more visible and more
easily scanned, hopefully reducing banner-blindness and making
the history notice more visible overall.

Note: we're only breaking the government name on larger sceen
sizes, to avoid orphan words and taking up too much space on
a smaller screen