Default to today for new (non historic) role appts
The common usecase is adding a new role appointment at the time
they appointment is made, and it makes more sense to set the
default started_at date to be today.

For historic role appointments, the default is left blank.

https://trello.com/c/7IVzcDP6/147-when-creating-new-roles-and-governments-set-the-default-date-to-today

Note: the selectors to test a date field are horrible and
would be worth refactoring in the future, when there are
more examples that a common pattern can be made form